Refactoring 3d example to new architecture
