+++ BitEngine/src/BitEngine/Core/ECS/RenderableMeshComponent.h
@@ -16,11 +16,11 @@ namespace BitEngine{
                 : mesh(nullptr), material(nullptr)
             {}
 
-            RenderableMeshComponent(const Mesh* _mesh, Material* _material)
+            RenderableMeshComponent(Mesh* _mesh, Material* _material)
                 : mesh(_mesh), material(_material)
             {}
 
-            void setMesh(const Mesh* _mesh, Material* _mat = nullptr){
+            void setMesh(Mesh* _mesh, Material* _mat = nullptr){
                 if (_mat == nullptr)
                     _mat = _mesh->getMaterial(); // Set default material
 
@@ -28,7 +28,7 @@ namespace BitEngine{
                 material = _mat;
             }
 
-            const Mesh* getMesh() const {
+            Mesh* getMesh() const {
                 return mesh;
             }
 
@@ -38,7 +38,7 @@ namespace BitEngine{
 
         private:
 
-            const Mesh* mesh;
+            Mesh* mesh;
             Material* material; // Override mesh's default material
 
 	};
