+++ samples/sample01/src/Game/Common/GameGlobal.h
